backend/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.11.2
certifi==2022.9.24
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
confluent-kafka==2.3.0
cryptography==41.0.5
distlib==0.3.7
Django==4.2.7
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==4.3.0
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
filelock==3.13.1
Flask==2.2.2
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
MarkupSafe==2.1.2
MouseInfo==0.1.3
Pillow==9.4.0
platformdirs==3.11.0
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyMySQL==1.1.0
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pytweening==1.0.4
pytz==2023.3.post1
pywhatkit==5.4
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.12
virtualenv==20.24.6
virtualenvwrapper-win==1.2.7
Werkzeug==2.2.2
wikipedia==1.4.0
zipp==3.13.0
